requirements.txt
torch>=2.1
numpy>=1.24
Pillow>=10.0
tqdm>=4.66
PyYAML>=6.0